pyproject.toml
[tool.poetry]
name = "ah-cara"
version = "0.1.0"
description = ""
authors = ["Paulo Leitão <paulojbleitao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
